General code improvement
