+++ src/me/wheezygold/skripthack/SkriptHack.java
@@ -1,11 +1,10 @@
 package me.wheezygold.skripthack;
 
 import ch.njol.skript.Skript;
+import me.wheezygold.skripthack.commands.Skripthack;
+import me.wheezygold.skripthack.util.Metrics;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 import org.bukkit.event.Listener;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -20,6 +19,7 @@ public class SkriptHack extends JavaPlugin implements Listener {
     public void onEnable() {
         instance = this;
         if (getServer().getPluginManager().getPlugin("Skript")!=null) {
+            this.getCommand("skript-hack").setExecutor(new Skripthack());
             log(C.cAqua + "#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#+#");
             log(C.cAqua + "This plugin's use is at your own risk!");
             log(C.cAqua + "This plugin can stop working at any moment and is not supported by Skript.");
@@ -43,7 +43,7 @@ public class SkriptHack extends JavaPlugin implements Listener {
                 log("Anyway, I'll just hack my own syntax in.");
                 try {
                     hackSkript(true);
-                    Skript.registerAddon(this).loadClasses("me.wheezygold.skripthack", "skript");
+                    Skript.registerAddon(this).loadClasses("me.wheezygold.skripthack", "syntax");
                     hackSkript(false);
                     log("Loaded Syntax (Hacked Mode)");
                 } catch (IOException | NoSuchFieldException | IllegalAccessException e) {
@@ -60,58 +60,15 @@ public class SkriptHack extends JavaPlugin implements Listener {
     public void onDisable() {
 
     }
-
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (command.getName().equalsIgnoreCase("skript-hack")) {
-            if (args.length >= 1) {
-                if (sender instanceof Player) {
-                    final Player p = (Player) sender;
-                    if (args[0].equalsIgnoreCase("true")) {
-                        p.sendMessage(ChatColor.BLUE + "skript-hack >> " + ChatColor.GRAY + "Skript's registration system has been unlocked.");
-                        try {
-                            hackSkript(true);
-                        } catch (NoSuchFieldException | IllegalAccessException e) {
-                            e.printStackTrace();
-                        }
-                        return true;
-                    } else if (args[0].equalsIgnoreCase("false")) {
-                        p.sendMessage(ChatColor.BLUE + "skript-hack >> " + ChatColor.GRAY + "Skript's registration system has been locked.");
-                        try {
-                            hackSkript(false);
-                        } catch (NoSuchFieldException | IllegalAccessException e) {
-                            e.printStackTrace();
-                        }
-                        return true;
-                    } else {
-                        return false;
-                    }
-
-                } else {
-                    if (args[0].equalsIgnoreCase("true")) {
-                        System.out.println(ChatColor.BLUE + "skript-hack >> " + ChatColor.GRAY + "Skript's registration system has been unlocked.");
-                        return true;
-                    } else if (args[0].equalsIgnoreCase("false")) {
-                        System.out.println(ChatColor.BLUE + "skript-hack >> " + ChatColor.GRAY + "Skript's registration system has been locked.");
-                        return true;
-                    } else {
-                        return false;
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    private void log(String message) {
-        getServer().getConsoleSender().sendMessage("[skript-hack] " + message);
+    private static void log(String message) {
+        instance.getServer().getConsoleSender().sendMessage("[skript-hack] " + message);
     }
 
     public static SkriptHack getInstance() {
         return instance;
     }
 
-    public void hackSkript(boolean allowReg) throws NoSuchFieldException, IllegalAccessException {
+    public static void hackSkript(boolean allowReg) throws NoSuchFieldException, IllegalAccessException {
         Field field;
         field = Skript.class.getDeclaredField("acceptRegistrations");
         field.setAccessible(true);
