+++ src/me/wheezygold/skripthack/syntax/effects/EffOpenHack.java
@@ -1,4 +1,4 @@
-package me.wheezygold.skripthack.skript;
+package me.wheezygold.skripthack.syntax.effects;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.lang.Effect;
@@ -6,8 +6,8 @@ import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser;
 import ch.njol.skript.util.Timespan;
 import ch.njol.util.Kleenean;
-import me.wheezygold.skripthack.SKU;
 import me.wheezygold.skripthack.SkriptHack;
+import me.wheezygold.skripthack.util.SKU;
 import org.bukkit.event.Event;
 
 @SuppressWarnings("unused")
